src/main/webapp/app/entities/gift/route/gift-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IGift, Gift } from '../gift.model';
import { GiftService } from '../service/gift.service';

import { GiftRoutingResolveService } from './gift-routing-resolve.service';

describe('Gift routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: GiftRoutingResolveService;
  let service: GiftService;
  let resultGift: IGift | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(GiftRoutingResolveService);
    service = TestBed.inject(GiftService);
    resultGift = undefined;
  });

  describe('resolve', () => {
    it('should return IGift returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGift = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGift).toEqual({ id: 123 });
    });

    it('should return new IGift if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGift = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultGift).toEqual(new Gift());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Gift })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGift = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGift).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
